+++ e2e/nx-verdaccio-e2e/test/plugin-create-nodes.e2e.test.ts
@@ -190,6 +190,7 @@ describe('nx-verdaccio plugin create-nodes-v2', () => {
     expect(projectJson.targets).toStrictEqual(
       expect.objectContaining({
         e2e: expect.objectContaining({
+          configurations: {},
           dependsOn: [
             {
               params: 'forward',
@@ -199,9 +200,6 @@ describe('nx-verdaccio plugin create-nodes-v2', () => {
         }),
         [TARGET_ENVIRONMENT_BOOTSTRAP]: expect.objectContaining({
           executor: '@push-based/nx-verdaccio:env-bootstrap',
-          options: {
-            environmentRoot: 'tmp/environments/lib-a-e2e',
-          },
         }),
         [TARGET_ENVIRONMENT_INSTALL]: expect.objectContaining({
           dependsOn: [
@@ -211,15 +209,18 @@ describe('nx-verdaccio plugin create-nodes-v2', () => {
               target: TARGET_PACKAGE_INSTALL,
             },
           ],
-          executor: 'nx:noop',
-          options: { environmentRoot: 'tmp/environments/lib-a-e2e' },
+          executor: 'nx:run-commands',
+          options: {
+            environmentRoot: expect.toMatchPath('tmp/environments/lib-a-e2e'),
+            command: expect.stringContaining(
+              'echo "dependencies installed for'
+            ),
+          },
         }),
         [TARGET_ENVIRONMENT_SETUP]: expect.objectContaining({
           cache: true,
           executor: '@push-based/nx-verdaccio:env-setup',
-          options: {
-            environmentRoot: 'tmp/environments/lib-a-e2e',
-          },
+          options: {},
           inputs: [
             '{projectRoot}/project.json',
             {
@@ -245,21 +246,28 @@ describe('nx-verdaccio plugin create-nodes-v2', () => {
           options: expect.objectContaining({
             clear: true,
             config: '.verdaccio/config.yml',
-            environmentDir: 'tmp/environments/lib-a-e2e',
+            environmentDir: expect.toMatchPath('tmp/environments/lib-a-e2e'),
             port: expect.any(Number), // random port number
             projectName: 'lib-a-e2e',
-            storage: 'tmp/environments/lib-a-e2e/storage',
+            storage: expect.toMatchPath('tmp/environments/lib-a-e2e/storage'),
           }),
         }),
         [TARGET_ENVIRONMENT_VERDACCIO_STOP]: expect.objectContaining({
           executor: '@push-based/nx-verdaccio:kill-process',
           options: {
-            filePath: 'tmp/environments/verdaccio-registry.json',
+            filePath: expect.toMatchPath(
+              'tmp/environments/verdaccio-registry.json'
+            ),
           },
         }),
         [TARGET_ENVIRONMENT_E2E]: expect.objectContaining({
           executor: '@push-based/nx-verdaccio:env-teardown',
-          dependsOn: ['e2e'],
+          dependsOn: [
+            {
+              params: 'forward',
+              target: 'e2e',
+            },
+          ],
         }),
         [TARGET_ENVIRONMENT_TEARDOWN]: expect.objectContaining({
           executor: '@push-based/nx-verdaccio:env-teardown',
@@ -267,16 +275,6 @@ describe('nx-verdaccio plugin create-nodes-v2', () => {
       })
     );
 
-    expect({
-      ...projectJson.targets,
-      [TARGET_ENVIRONMENT_VERDACCIO_START]: {
-        ...projectJson.targets?.[TARGET_ENVIRONMENT_VERDACCIO_START],
-        options: {
-          ...projectJson.targets?.[TARGET_ENVIRONMENT_VERDACCIO_START].options,
-          port: expect.any(Number),
-        },
-      },
-    }).toMatchSnapshot();
   });
 
   it('should NOT add environment targets to project without targetName e2e', async () => {