+++ examples/e2e/cli-static-e2e/test/cli-command-sort.spec.ts
@@ -1,28 +1,22 @@
-import { dirname, join, basename } from 'node:path';
+import { basename, dirname, join } from 'node:path';
 import { afterEach, describe, expect, it } from 'vitest';
-import { mkdir, readFile, rm, writeFile } from 'node:fs/promises';
+import { readFile, rm } from 'node:fs/promises';
 import { executeProcess, objectToCliArgs } from '@push-based/test-utils';
-import {
-  getTestFixturesDist,
-  getTestEnvironmentRoot,
-} from '@push-based/test-utils';
+import type { SimpleGit } from 'simple-git';
+import { simpleGit } from 'simple-git';
 
 describe('CLI command - sort', () => {
-  const fixturesDist = getTestFixturesDist('cli-command-sort', {
-    root: getTestEnvironmentRoot(process.env['NX_TASK_TARGET_PROJECT']),
-  });
+  const envRoot = join('static-environments', 'user-lists');
+  const baseDir = join(envRoot, 'src', 'lib');
+  const gitClient: SimpleGit = simpleGit(process.cwd());
 
   afterEach(async () => {
-    await rm(fixturesDist, { recursive: true, force: true });
+    await gitClient.checkout([baseDir]);
+    await gitClient.clean('f', [baseDir]);
   });
 
   it('should execute CLI command sort when param file is given', async () => {
-    const testPath = join(fixturesDist, 'execute-sort-command', 'users.json');
-    await mkdir(dirname(testPath), { recursive: true });
-    await writeFile(
-      testPath,
-      JSON.stringify([{ name: 'Michael' }, { name: 'Alice' }])
-    );
+    const testPath = join(baseDir, 'unsorted-users.json');
 
     const { code } = await executeProcess({
       command: 'npx',
@@ -37,7 +31,34 @@ describe('CLI command - sort', () => {
     expect(code).toBe(0);
 
     const content = (await readFile(testPath)).toString();
-    expect(JSON.parse(content)).toEqual([
+    expect(JSON.parse(content)).toStrictEqual([
+      { name: 'Alice' },
+      { name: 'Michael' },
+    ]);
+  });
+
+  it('should execute CLI command sort on sorted file', async () => {
+    const testPath = join(baseDir, 'sorted-users.json');
+
+    const contentBefore = (await readFile(testPath)).toString();
+    expect(JSON.parse(contentBefore)).toStrictEqual([
+      { name: 'Alice' },
+      { name: 'Michael' },
+    ]);
+
+    const { code } = await executeProcess({
+      command: 'npx',
+      args: objectToCliArgs({
+        _: ['cli', 'sort'],
+        filePath: basename(testPath),
+      }),
+      cwd: baseDir,
+    });
+
+    expect(code).toBe(0);
+
+    const content = (await readFile(testPath)).toString();
+    expect(JSON.parse(content)).toStrictEqual([
       { name: 'Alice' },
       { name: 'Michael' },
     ]);