+++ examples/projects/cli/src/lib/cli.ts
@@ -3,6 +3,7 @@ import { sortUserFile } from '@push-based/core';
 
 export type CliArgs = {
   filePath: string;
+  plugin?: string;
 };
 
 const NOOP_BUILDER = <T>(_: Argv<T>) => void 0;
@@ -17,13 +18,26 @@ export function cli(args: string[]) {
         description: 'Path to the user file',
         demandOption: true,
       },
+      plugin: {
+        type: 'string',
+        description: 'Path to the user file',
+        alias: 'p',
+      },
     } satisfies Record<keyof CliArgs, Options>)
     .command('*', 'Sort users', NOOP_BUILDER, sortCommandHandle)
     .command('sort', 'Sort users', NOOP_BUILDER, sortCommandHandle);
 }
 
 export async function sortCommandHandle(args: CliArgs) {
-  const { filePath } = args;
-  await sortUserFile(filePath);
-  console.log(`Sorted users in ${filePath}`);
+  const { filePath, plugin } = args;
+  try {
+    const sortFn = plugin
+      ? await import(plugin).then(({ default: d }) => d)
+      : sortUserFile;
+    await sortFn(filePath);
+    console.log(`Sorted users in ${filePath}`);
+  } catch (error) {
+    console.error(`Failed to load plugin: ${plugin}`);
+    throw error;
+  }
 }