+++ projects/nx-verdaccio/src/executors/env-bootstrap/schema.ts
@@ -1,6 +1,9 @@
-export type BootstrapExecutorOptions = Partial<{
-  keepServerRunning: boolean;
-  environmentRoot: string;
-  printConfig: boolean;
-  verbose: boolean;
-}>;
+import { type Environment } from './npm';
+
+export type BootstrapExecutorOptions = Partial<
+  {
+    keepServerRunning: boolean;
+    printConfig: boolean;
+    verbose: boolean;
+  } & Environment
+>;