+++ projects/nx-verdaccio/src/executors/env-teardown/teardown-env.ts
@@ -7,29 +7,16 @@ import { VERDACCIO_REGISTRY_JSON } from '../env-bootstrap/constants';
 import { fileExists } from '../../internal/file-system';
 import { rm } from 'node:fs/promises';
 import runKillProcessExecutor from '../kill-process/executor';
-import { DEFAULT_ENVIRONMENTS_OUTPUT_DIR } from '../../plugin/constants';
-import { type ExpandedPluginConfiguration } from 'nx/src/config/nx-json';
-import type { NxVerdaccioCreateNodeOptions } from '../../plugin/schema';
+import { getEnvironmentRoot } from '../../internal/environment-root';
 
 export const gitClient: SimpleGit = simpleGit(process.cwd());
 export type TeardownEnvironmentOptions = Environment & { verbose?: boolean };
 
 export async function teardownEnvironment(
   context: ExecutorContext,
-  options: TeardownEnvironmentOptions,
-  git: SimpleGit = gitClient
+  options: TeardownEnvironmentOptions
 ): Promise<void> {
-  const { environmentRoot: optEnvironmentRoot } = options;
-  const plugin = context.nxJsonConfiguration.plugins.find((pCfg) => {
-    return (
-      typeof pCfg === 'object' && pCfg?.plugin === '@push-based/nx-verdaccio'
-    );
-  }) as ExpandedPluginConfiguration<NxVerdaccioCreateNodeOptions>;
-  const environmentsDir =
-    plugin.options.environments?.environmentsDir ??
-    DEFAULT_ENVIRONMENTS_OUTPUT_DIR;
-  const environmentRoot =
-    optEnvironmentRoot ?? join(environmentsDir, context.projectName);
+  const environmentRoot = getEnvironmentRoot(context, options);
   // kill verdaccio process if running
   const registryPath = join(environmentRoot, VERDACCIO_REGISTRY_JSON);
   const registryJsonExists = await fileExists(registryPath);
@@ -46,9 +33,24 @@ export async function teardownEnvironment(
 
   const environmentRootInRepo = await isFolderInGit(environmentRoot);
   if (environmentRootInRepo) {
-    // await git.checkout([environmentRoot]);
-    // await git.clean('f', [environmentRoot]);
+    await gitClient.checkout([environmentRoot]);
+    await gitClient.clean('f', [environmentRoot]);
     logger.info(`Cleaned git history in ${environmentRoot}`);
+
+    const nodeModules = join(environmentRoot, 'node_modules');
+    try {
+      await rm(nodeModules, {
+        recursive: true,
+        force: true,
+        retryDelay: 100,
+        maxRetries: 2,
+      });
+      logger.info(`deleted folder ${nodeModules}.`);
+    } catch (error) {
+      throw new Error(
+        `Error cleaning history of folder ${nodeModules}. ${error.message}`
+      );
+    }
   } else {
     try {
       await rm(environmentRoot, {