+++ projects/nx-verdaccio/src/executors/pkg-install/executor.unit-test.ts
@@ -68,10 +68,7 @@ describe('runNpmInstallExecutor', () => {
 
     expect(logger.info).toHaveBeenCalledTimes(1);
     expect(logger.info).toHaveBeenCalledWith(
-      `Installing my-lib@1.0.0 in ${join(
-        DEFAULT_ENVIRONMENTS_OUTPUT_DIR,
-        'my-lib-e2e'
-      )}`
+      expect.stringContaining(`Installing my-lib@1.0.0 in`)
     );
 
     expect(executeProcessSpy).toHaveBeenCalledTimes(1);
@@ -81,11 +78,16 @@ describe('runNpmInstallExecutor', () => {
         args: [
           'install',
           'my-lib@1.0.0',
+          '--include=prod',
+          // @TODO implement options
+          //'--include=dev',
+          //'--include=optional',
+          '--include=peer',
           '--no-fund',
           '--no-shrinkwrap',
           '--save',
         ],
-        cwd: 'tmp/environments/my-lib-e2e',
+        cwd: expect.toMatchPath('tmp/environments/my-lib-e2e'),
         verbose: true,
       })
     );