+++ projects/nx-verdaccio/src/plugin/targets/environment.targets.ts
@@ -71,9 +71,7 @@ export function verdaccioTargets(
     [TARGET_ENVIRONMENT_VERDACCIO_START]: {
       // @TODO: consider using the executor function directly to reduce the number of targets
       // https://github.com/nrwl/nx/blob/b73f1e0e0002c55fc0bacaa1557140adb9eec8de/packages/js/src/executors/verdaccio/verdaccio.impl.ts#L22
-      outputs: [
-        `{options.environmentRoot}/${VERDACCIO_STORAGE_DIR}`,
-      ],
+      outputs: [`{options.environmentRoot}/${VERDACCIO_STORAGE_DIR}`],
       executor: '@nx/js:verdaccio',
       options: {
         config: '.verdaccio/config.yml',
@@ -113,7 +111,6 @@ export function getEnvTargets(
   return {
     [TARGET_ENVIRONMENT_BOOTSTRAP]: {
       executor: `${PACKAGE_NAME}:${EXECUTOR_ENVIRONMENT_BOOTSTRAP}`,
-      options: { environmentRoot },
     },
     // intermediate task just here to execute dependent pkg-install tasks with the correct environmentProject
     [TARGET_ENVIRONMENT_INSTALL]: {
@@ -126,7 +123,7 @@ export function getEnvTargets(
       ],
       options: { environmentRoot },
       // This is here to make it appear in the graph in older nx versions (otherwise it is filtered out)
-      command: `echo "dependencies installed for ${environmentRoot}"`
+      command: `echo "dependencies installed for ${environmentRoot}"`,
     },
     // runs env-bootstrap-env, install-env and stop-verdaccio
     [TARGET_ENVIRONMENT_SETUP]: {
@@ -155,9 +152,6 @@ export function getEnvTargets(
       ],
       cache: true,
       executor: `${PACKAGE_NAME}:${EXECUTOR_ENVIRONMENT_SETUP}`,
-      options: {
-        environmentRoot,
-      },
     },
     [TARGET_ENVIRONMENT_TEARDOWN]: {
       executor: `${PACKAGE_NAME}:${EXECUTOR_ENVIRONMENT_TEARDOWN}`,
@@ -165,7 +159,7 @@ export function getEnvTargets(
     [TARGET_ENVIRONMENT_E2E]: {
       dependsOn: targetNames.map((targetName) => ({
         target: targetName,
-        params: 'forward'
+        params: 'forward',
       })),
       executor: `${PACKAGE_NAME}:${EXECUTOR_ENVIRONMENT_TEARDOWN}`,
     },